refactor(sync): move storage ports below features
