+++ lib/core/db/repositories/api_config_repo.dart
@@ -4,7 +4,7 @@ import 'package:drift/drift.dart';
 
 import '../app_db.dart';
 import '../../models/api_config.dart';
-import '../../../features/cloud_sync/sync_repo_interfaces.dart';
+import '../../application/sync_repo_interfaces.dart';
 import '../../models/extra_request_parameter.dart';
 
 class ApiConfigRepo implements SyncApiConfigStore {
