+++ lib/core/db/repositories/lorebook_repo.dart
@@ -4,7 +4,7 @@ import 'package:drift/drift.dart';
 
 import '../app_db.dart';
 import '../../models/lorebook.dart';
-import '../../../features/cloud_sync/sync_repo_interfaces.dart';
+import '../../application/sync_repo_interfaces.dart';
 
 class LorebookRepo implements SyncLorebookStore {
   final AppDatabase _db;
