+++ lib/core/db/repositories/preset_repo.dart
@@ -4,7 +4,7 @@ import 'package:drift/drift.dart';
 
 import '../app_db.dart';
 import '../../models/preset.dart';
-import '../../../features/cloud_sync/sync_repo_interfaces.dart';
+import '../../application/sync_repo_interfaces.dart';
 
 class PresetRepo implements SyncPresetStore {
   final AppDatabase _db;
