+++ lib/core/db/repositories/studio_preset_repo.dart
@@ -2,7 +2,7 @@ import 'dart:convert';
 
 import 'package:drift/drift.dart';
 
-import '../../../features/cloud_sync/sync_repo_interfaces.dart';
+import '../../application/sync_repo_interfaces.dart';
 import '../../models/studio_config.dart';
 import '../app_db.dart';
