+++ lib/shared/application/sync_theme_store.dart
@@ -1,0 +1,6 @@
+import '../theme/theme_preset.dart';
+
+abstract class SyncThemePresetStore {
+  Future<List<ThemePreset>> getAll();
+  Future<void> putAll(List<ThemePreset> presets);
+}
